Update to new branch

+++ src/main/java/zarvis/bakery/models/Location.java
@@ -9,7 +9,7 @@ public class Location {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
@@ -17,7 +17,7 @@ public class Location {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
